ViewModels developed

+++ app/src/main/java/com/example/memory_app/domain/repository/Score.kt
@@ -1,4 +1,4 @@
 package com.example.memory_app.domain.repository
 
 @JvmInline
-value class Score(val value : Int)
+value class Score(val value : Double)
